Move StringDoubleMap
